move more code to net.adoptopenjdk.icedteaweb.resources
