+++ core/src/main/java/net/adoptopenjdk/icedteaweb/resources/CachedDaemonThreadPoolProvider.java
@@ -34,7 +34,7 @@
  this exception to your version of the library, but you are not
  obligated to do so.  If you do not wish to do so, delete this
  exception statement from your version. */
-package net.sourceforge.jnlp.cache;
+package net.adoptopenjdk.icedteaweb.resources;
 
 import java.util.concurrent.Executor;
 import java.util.concurrent.ExecutorService;
