+++ core/src/main/java/net/adoptopenjdk/icedteaweb/resources/UrlRequestResult.java
@@ -1,4 +1,4 @@
-package net.sourceforge.jnlp.cache;
+package net.adoptopenjdk.icedteaweb.resources;
 
 import net.adoptopenjdk.icedteaweb.jnlp.version.VersionId;
