+++ core/src/main/java/net/adoptopenjdk/icedteaweb/resources/downloader/ExactVersionedResourceDownloader.java
@@ -1,8 +1,8 @@
 package net.adoptopenjdk.icedteaweb.resources.downloader;
 
 import net.adoptopenjdk.icedteaweb.jnlp.version.VersionId;
+import net.adoptopenjdk.icedteaweb.resources.Resource;
 import net.adoptopenjdk.icedteaweb.resources.cache.Cache;
-import net.sourceforge.jnlp.cache.Resource;
 
 import java.net.URL;
 import java.util.List;
