+++ core/src/main/java/net/adoptopenjdk/icedteaweb/resources/initializer/ExactVersionedResourceInitializer.java
@@ -3,17 +3,17 @@ package net.adoptopenjdk.icedteaweb.resources.initializer;
 import net.adoptopenjdk.icedteaweb.jnlp.version.VersionId;
 import net.adoptopenjdk.icedteaweb.logging.Logger;
 import net.adoptopenjdk.icedteaweb.logging.LoggerFactory;
+import net.adoptopenjdk.icedteaweb.resources.Resource;
+import net.adoptopenjdk.icedteaweb.resources.ResourceUrlCreator;
 import net.adoptopenjdk.icedteaweb.resources.cache.Cache;
 import net.sourceforge.jnlp.DownloadOptions;
-import net.sourceforge.jnlp.cache.Resource;
-import net.sourceforge.jnlp.cache.ResourceUrlCreator;
 
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
 
-import static net.sourceforge.jnlp.cache.ResourceUrlCreator.getUrl;
-import static net.sourceforge.jnlp.cache.ResourceUrlCreator.getVersionedUrl;
+import static net.adoptopenjdk.icedteaweb.resources.ResourceUrlCreator.getUrl;
+import static net.adoptopenjdk.icedteaweb.resources.ResourceUrlCreator.getVersionedUrl;
 
 /**
  * Initializer for resources with an exact version.
