+++ core/src/main/java/net/sourceforge/jnlp/cache/NativeLibraryStorage.java
@@ -5,6 +5,7 @@ import net.adoptopenjdk.icedteaweb.io.FileUtils;
 import net.adoptopenjdk.icedteaweb.io.IOUtils;
 import net.adoptopenjdk.icedteaweb.logging.Logger;
 import net.adoptopenjdk.icedteaweb.logging.LoggerFactory;
+import net.adoptopenjdk.icedteaweb.resources.ResourceTracker;
 
 import java.io.File;
 import java.io.FileOutputStream;
