+++ core/src/test/java/net/adoptopenjdk/icedteaweb/resources/ResourceUrlCreatorTest.java
@@ -1,4 +1,4 @@
-package net.sourceforge.jnlp.cache;
+package net.adoptopenjdk.icedteaweb.resources;
 
 import net.adoptopenjdk.icedteaweb.http.HttpMethod;
 import net.adoptopenjdk.icedteaweb.jnlp.version.VersionString;
